perf: run the in-process sf commands in a worker thread so the extension host stays responsive at startup
