+++ src/worker.ts
@@ -1,5 +1,10 @@
 import { parentPort } from "worker_threads";
 import * as childProcess from "child_process";
+import {
+  loadSalesforceCoreFrom,
+  parseSfCommand,
+  runSfCoreCommand,
+} from "./utils/sfCoreCommands";
 
 // Execute a command concurrently
 
@@ -20,6 +25,56 @@ function execCliCommand(cmd: string, execOptions: any, requestId: string) {
   });
 }
 
+// --- In-process sf commands (@salesforce/core loaded from the installed CLI) ---
+// The library is loaded once per worker; every request then costs a few ms of
+// CPU here instead of a 2 to 4 s CLI process, and never blocks the extension host.
+
+const log = (message: string) => {
+  parentPort?.postMessage({ sfCoreLog: message });
+};
+
+let loadedCore: { coreDir: string; core: any } | null = null;
+
+function getCore(coreDir: string): any {
+  if (loadedCore && loadedCore.coreDir === coreDir) {
+    return loadedCore.core;
+  }
+  const t0 = Date.now();
+  const { core, version } = loadSalesforceCoreFrom(coreDir);
+  loadedCore = { coreDir, core };
+  log(
+    `Loaded @salesforce/core ${version} from ${coreDir} (${Date.now() - t0}ms, worker thread)`,
+  );
+  return core;
+}
+
+async function runSfCoreRequest(
+  request: { coreDir: string; command?: string; cwd?: string; preload?: boolean },
+  requestId: string | undefined,
+) {
+  try {
+    const core = getCore(request.coreDir);
+    if (request.preload || !request.command) {
+      return;
+    }
+    const parsed = parseSfCommand(request.command);
+    const result =
+      parsed === null
+        ? null
+        : await runSfCoreCommand(core, parsed, request.cwd, log);
+    parentPort?.postMessage({ requestId, sfCoreResult: result });
+  } catch (e: any) {
+    if (requestId) {
+      parentPort?.postMessage({
+        requestId,
+        sfCoreError: e?.message || String(e),
+      });
+    } else {
+      log(`preload failed: ${e?.message || String(e)}`);
+    }
+  }
+}
+
 if (parentPort) {
   parentPort.on("message", (msg) => {
     if (msg && msg.cliCommand) {
@@ -28,6 +83,8 @@ if (parentPort) {
         JSON.parse(msg.cliCommand.execOptions),
         msg.requestId,
       );
+    } else if (msg && msg.sfCore) {
+      void runSfCoreRequest(msg.sfCore, msg.requestId);
     }
   });
 }
